Seperated the utility functions and typsescript
