+++ src/components/ActivityCard/ActivityCard.tsx
@@ -4,38 +4,8 @@ import Link from 'next/link';
 import Image from 'next/image';
 import { BsBookmark } from 'react-icons/bs';
 import { useLocation } from '@/context/LocationContext';
-import { ActivityCardProps } from '@/types';
-
-const SUBCATEGORY_CLASSES: Record<string, string> = {
-  Bouldering: 'border-blue-500 text-blue-500',
-  Golfing: 'border-green-500 text-green-500',
-  Adventure: 'border-yellow-500 text-yellow-500',
-  Bowling: 'border-red-500 text-red-500',
-  Default: 'border-gray-300 text-gray-700',
-};
-
-const getSubCategoryClasses = (subCategory: string) =>
-  SUBCATEGORY_CLASSES[subCategory] || SUBCATEGORY_CLASSES.Default;
-
-// Utility to calculate distance using the Haversine formula
-const calculateDistance = (lat1: number, lon1: number, lat2: number, lon2: number): number => {
-  const R = 3958.8; // Radius of Earth in miles
-  const dLat = ((lat2 - lat1) * Math.PI) / 180;
-  const dLon = ((lon2 - lon1) * Math.PI) / 180;
-  const a =
-    Math.sin(dLat / 2) ** 2 +
-    Math.cos((lat1 * Math.PI) / 180) * Math.cos((lat2 * Math.PI) / 180) * Math.sin(dLon / 2) ** 2;
-  const c = 2 * Math.atan2(Math.sqrt(a), Math.sqrt(1 - a));
-  return R * c; // Distance in miles
-};
-
-const isValidCoordinate = ({
-  latitude,
-  longitude,
-}: {
-  latitude: number;
-  longitude: number;
-}): boolean => latitude !== 0 && longitude !== 0;
+import { ActivityCardProps } from '@/types/activities';
+import { getSubCategoryClasses, calculateDistance, isValidCoordinate } from '@/utils/activityCard';
 
 function ActivityCard({
   activity_id,
@@ -69,6 +39,7 @@ function ActivityCard({
             width={500}
             height={400}
             className="h-48 w-full rounded-2xl object-cover"
+            priority
           />
           <div className="absolute right-3 top-3 text-white opacity-0 transition-opacity duration-300 group-hover:opacity-100">
             <BsBookmark size={24} />
