+++ tailwind.config.ts
@@ -4,6 +4,8 @@ export default {
   content: [
     './src/pages/**/*.{js,ts,jsx,tsx,mdx}',
     './src/components/**/*.{js,ts,jsx,tsx,mdx}',
+    './src/utils/**/*.{js,ts,jsx,tsx,mdx}',
+
     './src/app/**/*.{js,ts,jsx,tsx,mdx}',
   ],
   theme: {
